Move variables config out of scss dir
