+++ client/config/gutter-width.js
